Adopt library mouse scroll behaviour for thumbnail
To keep it consistent with the library, scrolling with the mouse in
thumbnail mode now also changes the selection instead of just scrolling
the view.
